createPostBody progressing
Trying to make the main function both more logical and more readable by
splitting this stuff out of it. It's sort of becoming a director. Here
I'm writing createPostBody so that it can create both sections (new
projects and ending soon) in the markdown file depending on how it's
called. It's job should be just to write data out to a file, that's it.