More main function args and validation
Because I needed to test several arguments in the same way, I wrote a
simple helper function to do that.

The new args - newWindow and endWindow - were added to make the window
of time I'm collecting data for part of the function call. By assigning
the args defaults, I get the same effect as hard-coding a variable with
better code.