Fixed several bugs
1. Bug that resulted in "NA" being filled in for all fields. The hack
that relied on specific line numbers to locate where information was
found was, indeed, a hack. If the creator inserted a carriage return in
the wrong spot the data was lost. I learned how to use grep, made a
helper function that can help extract the data, and this should be a
much more stable solution.

2. Dates. *sigh*
Dates are always a pain in the ass and the way these are formatted here
is no different. Again, learning grep has helped make this string
manipulation a lot more legible. Plus, for whatever reason, suddenly end
dates started having the year crossover problem that start dates had, so
I made a similar helper function to ensure that years were captured
correctly in all cases.