Simplified createKsPost function call and added logMessage
A side effect of using sink() to write output is that log messages have to go somewhere else, so, replacing all the print messages will be a thing. Also, trying to make this more scriptlike and have the only variable in the main function call be the thing that actually routinely varies.